+++ CHANGELOG.md
@@ -1,15 +1,20 @@
-v0.3.0
+0.4.0
+------
+
+- Add pretty printing for union cases using overloaded `ToString` method
+
+0.3.0
 ------
 
 - Add generic support for unions
 - Implement code generation using nullable reference types feature
 
-v0.2.0
+0.2.0
 ------
 
 - Add equality comparison for unions
 
-v0.1.0
+0.1.0
 ------
 
 - Initial version